Updating every 24h
Activities.json erases itself every 24 hours and all the data will be saved to the large_json.json.

This is important because when I visualize the data, I do not want to see every single time entry for every tab ive been on. The large_json.json contains all the useful information of how long I spent on which app. Later on, if I feel like it, I can make sure that the Chrome key also has individual websites on the large_json.json (activities.json already has that) to track more efficiently.